python-services/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
groq>=1.1.2
cloudinary==1.41.0
moviepy==1.0.3
requests==2.32.3
python-multipart==0.0.9
aiofiles==24.1.0
yt-dlp==2024.9.27
Pillow==10.4.0
numpy==2.1.0
imageio==2.35.1
imageio-ffmpeg==0.5.1
decorator==4.4.2
tqdm==4.66.5
openai-whisper
setuptools-rust